requirements.txt
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
python-dateutil==2.8.2
django-cors-headers==3.14.0
djoser==2.2.0
Pillow==9.5.0